Tweaks to get cancellation txs building
First, script hashes often come "double-encoded"; up until Lucid 0.9.0,
this was the format that was expected; Upgrading to 0.9.0 fixed this
issue.

Second, the signerKey was being pulled from the *address* of the UTXO
passed in; this is just the escrow script, and we'd never be able to
produce a signature for that; additionally, the script execution sees it
as the wrong person trying to cancel the transaction, and so fails.

We'll need to either pass the expected signer in, or parse it from the
datum; We'll also need to consider the case where you want to cancel as
the alternate.

This produces a working transaction; Calvin can clean up the rest.