chore: update fee calculation
Fee calculation was tied to a completed transaction, but this caused issues in the past when the transaction wasn't ready to be signed. When we tried to complete it again, WASM memory references were no longer valid as they are discarded upon completion. This update creates a cloned transaction during the initial complete phase that calculates transaction fee and CBOR encoding of the transaction before it is signed, while also keeping the real reference to be signed later.