chore: isolate txbuilder params
Contract parameters were kept in SundaeUtils. This was insufficient for scaling contract types, so this PR moves them to their respective TxBuilder extensions and includes tests for their values. It also reduces the max scooper fee for the V3 TxBuilder to 1 ADA, instead of 1.1. It also adds a pipeline for running tests on PRs